tests/requirements.txt
# Python API Test Framework Dependencies
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.3.0
aiohttp>=3.8.0
faker>=19.0.0
python-dotenv>=1.0.0
pymongo>=4.5.0
colorama>=0.4.6
tabulate>=0.9.0
